// ui-packages/ui-components/lib/sidenav/sidenav.type.ts
export enum SidenavExamplesIds {
  Device = 'Device',
  BuiltIn = 'BuiltIn',
  FromLibraries = 'FromLibraries',
}

export enum SidenavLibrariesIds {
  Standard = 'Standard',
  Custom = 'Custom',
  Favorites = 'Favorites',
}

export enum SidenavReferenceIds {
  Functions = 'Functions',
  Variables = 'Variables',
  Structure = 'Structure',
}

export type SidenavTabsIds =
  | SidenavExamplesIds
  | SidenavLibrariesIds
  | SidenavReferenceIds;

export type SidenavTabItem =
  | {
      id: SidenavExamplesIds;
      label: MessageDescriptor;
    }
  | {
      id: SidenavLibrariesIds;
      label: MessageDescriptor;
    }
  | {
      id: SidenavReferenceIds;
      label: MessageDescriptor;
    };

export type SidenavTabAction = () => void;

export type SidenavExamplesDictionary<T> = { [K in SidenavExamplesIds]: T };
export type SidenavLibrariesDictionary<T> = { [K in SidenavLibrariesIds]: T };
export type SidenavReferenceDictionary<T> = { [K in SidenavReferenceIds]: T };

export type SidenavExamplesItemDictionary =
  SidenavExamplesDictionary<SidenavTabItem>;
export type SidenavExamplesHandlerDictionary =
  SidenavExamplesDictionary<SidenavTabAction>;

export type SidenavLibrariesItemDictionary =
  SidenavLibrariesDictionary<SidenavTabItem>;
export type SidenavLibrariesHandlerDictionary =
  SidenavLibrariesDictionary<SidenavTabAction>;

export type SidenavReferenceItemDictionary =
  SidenavReferenceDictionary<SidenavTabItem>;

export enum ExamplesMenuItemIds {
  CopyInYourSketches = 'CopyInYourSketches',
  Download = 'Download',
}
// ...
export type ExamplesMenuItemType = DropdownMenuItemType<
  ExamplesMenuItemIds,
  MessageDescriptor
>;
export type ExamplesMenuSection = DropdownMenuSectionType<
  ExamplesMenuItemIds,
  MessageDescriptor
>;

export type ExamplesMenuDictionary<T> = { [K in ExamplesMenuItemIds]: T };
export type ExamplesMenuItemDictionary =
  ExamplesMenuDictionary<ExamplesMenuItemType>;